refactor and cleanup - define .graphql within each resource (e.g., movie.graphql inside /movie) - adjust GraphQLModule config
